pyproject.toml
[tool.poetry]
name = "fastapidemo"
version = "0.1.0"
description = ""
authors = ["peihsuanlin5566 <jnes40722@gmail.com>"]
readme = "README.md"
packages = [ {include = "api"}, {include = "api_v2"} ]


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.65.1"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
sqlalchemy = "^1.4.45"
aiomysql = "^0.1.1"
greenlet = "^2.0.1"
snowflake-sqlalchemy = "^1.4.4"
numpy = "^1.24.1"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.3"
aiosqlite = "^0.18.0"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
